power: supply: qti: Disable notifications while in sleep
Currently during fast charging, charger FW on ADSP sends battery power
supply notifications to HLOS even while in sleep. These notifications
cause HLOS to come out of sleep to process the request. This is not needed
as display is OFF and any change in battery status will not be updated on
screen.

Add panel event notifiers to receive notifications from display FW so that
power supply notifications from the charger FW can be disabled when the
panel is turned off and re-enabled when the panel is on.

Change-Id: Ibf9e688f1571fd40ca18a1155fce95efd039690b
Signed-off-by: Anjelique Melendez <quic_amelende@quicinc.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
[ghostrider-reborn: adapt to mi disp notifier]
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>